Refactor resource type and filename conversion
